requirements.txt
django==4.2
djangorestframework==3.14
mysqlclient==2.1.1
python-dotenv==1.0.0
djangorestframework-simplejwt==5.1.0
Pillow
django-cors-headers